+++ Hoteleria/commons/src/main/java/com/tcg/commons/dto/HabitacionResponse.java
@@ -1,11 +1,13 @@
 package com.tcg.commons.dto;
 
+import java.math.BigDecimal;
+
 public record HabitacionResponse(
-	    Long id,
-	    Integer numero,
-	    String tipo,
-	    String descripcion,
-	    Double precio,
-	    Integer capacidad,
-	    String estado
-	) {}
+    Long id,
+    Integer numero,
+    String tipo,
+    String descripcion,
+    BigDecimal precio,
+    Integer capacidad,
+    String estado
+) {}
